requirements.txt
youtube_dl==2020.6.16.1
pyNaCl
discord.py[voice]